Refactor: Split Facade into Interface and Implementation
